TextSelection: improve logic that sync handles with selection
Since property 'selectionPosition' that depends on 'selectionStart'
gets updated after the call to 'onSelectionStartChanged', we cannot
depend on it inside the handler. Moreover, since selectionStart
gets updated before cursorPosition when selecting a new selection, we
need to update both selectionHandle and cursorHandle in both cases
to ensure that we base the logic on the latest values.

This will fix two bugs that causes the handles to jump to wrong
positions (and even wrong lines) once the user starts dragging.

Change-Id: I74919abb4169439659bd42952d5d5e6020f33853