TextHandle: fix mapping of touch position to handle position
Regardless of where the user drags on the mouse area (which
can be big for the user to hit it), you should continue dragging
on the selection where it is, not jump the location under the finger.

This patch will change offset to be the distance from the text position
under the handle to the touch point, so regardless where the user
touches the handle, the selection will not jump when he starts to drag.

Change-Id: I849e42c5e94f44aec355224a9db146bcdf22d43f