TextHandle: change when a handle should be enabled and visible
Only let a handle be enabled if it's visible, otherwise the
handle will still be in the scene and eat mouse events.
The result will be that you e.g cannot move the cursor to
the place where you left the handle, since you cannot touch
the input field underneath.
And keep the handles visible when there is a selection, and hide them
otherwise. No need for extra calculation logic that was needed
back when the handles got reparented out the flickable.

Change-Id: I2147824629c321096ebadf6252e41a88766dc527