TextHandle: ensure handles end up aligned with cursor rect
Since the handles are explixitly reparented higher up in the
hierarchy, we need to map the cursor rect of the edit to the
parent, rather that trying to do the calculations manually.
Otherwise we end up not taking all padding and margins into
account. We also need to align the text selection with the
physical position of the text, which is itself pixel-aligned.

Change-Id: I279fb69ffa18d58bf584aeafee0a84865364fcca