TextArea: don't reparent handles out of flickable
Reparating them higher up sounds like a good idea, but it
causes strange drawing artifacts other places. E.g you would
need to ensure that the handels follows the text area pixel
perfect while flicking, which is hard to track and needs hacks to work.
Moreover, the handles would also be drawn when selection is flicked outside
the view (since they would follow the coordinates of the content view), which
will make them appear on top of other components. In the end, just keeping
them as children of the text area works the best, even if they get clipped in some
cases. This is also how native controls on e.g iOS does it.

Change-Id: Ie119c3a725dd0fc451fba83e390e2cbc93685cf3