TextHandle: add clicked signal
Add a clicked signal to TextHandle that emits
if the handle is clicked without being dragged
Since the handle will move along with the drag, we
cannot rely on the MouseAreas own clicked signal
since that one will get confused and fire also after
a drag.

The handle instances will use this signal to move the
cursor where the user clicked, even when the user
clicks on a handle.

Change-Id: I4e611e73a6f76822e1008ccf75f0f5562fc673dc